move translate-api.js to utilities folder
