+++ messages/message.controller.js
@@ -1,5 +1,5 @@
 import fs from "fs"
-import { translator } from "../translate-api.js"
+import { translator } from "../utilities/translate-api.js"
 
 export function getMessage(req, res) {
     let existingMessage = JSON.parse(fs.readFileSync('./messages/messages.json'), 'utf8')
@@ -15,7 +15,6 @@ export async function createMessage(req, res) {
     console.log(existingMessage)
     const message = req.body;
     const translatedText = await translator(`Please translate ${message.text} from ${message.sourceLanguage} to ${message.targetLanguage}`)
-    console.log(translatedText)
     const formattedMessage = {
         [message.user_email]: [{
             user_id: message.user_email,
