+++ messages/messages.json
@@ -15,6 +15,14 @@
             "targetLanguage": "mandarin",
             "translation": "嗨，你最喜欢的歌曲是什么？ (Hāi, nǐ zuì xǐhuan de gēqǔ shì shénme?)",
             "createdAt": 1701834126472
+        },
+        {
+            "user_id": "234@123.com",
+            "text": "hi, what's your favorite film?",
+            "sourceLanguage": "english",
+            "targetLanguage": "mandarin",
+            "translation": "你好，你最喜欢的电影是什么？ (Nǐ hǎo, nǐ zuì xǐhuān de diànyǐng shì shénme?)",
+            "createdAt": 1701834381662
         }
     ]
 }
